+++ test/PriorityQueue.test.js
@@ -129,6 +129,40 @@ describe('PriorityQueue', () => {
     });
   });
 
+  describe('fix', () => {
+    it('fixes element positions when multiple priorities change', () => {
+      const one = { id: 'one', priority: 1 };
+      const two = { id: 'two', priority: 2 };
+      const three = { id: 'three', priority: 3 };
+      const four = { id: 'four', priority: 4 };
+      const five = { id: 'five', priority: 5 };
+      const six = { id: 'six', priority: 6 };
+      const seven = { id: 'seven', priority: 7 };
+      const qTest = PriorityQueue.fromArray(
+        [one, two, three, four, five, six, seven],
+        (a, b) => a.priority - b.priority
+      );
+
+      one.priority = 8;
+      four.priority = 0;
+      seven.priority = 3.5;
+
+      expect(qTest.fix()).to.equal(qTest);
+      expect(qTest.size()).to.equal(7);
+      expect(qTest.toArray()).to.eql([
+        four,
+        two,
+        three,
+        seven,
+        five,
+        six,
+        one
+      ]);
+      expect(qTest.front()).to.equal(four);
+      expect(qTest.back()).to.equal(one);
+    });
+  });
+
   describe('iterator', () => {
     it('allows iterating on queue elements', () => {
       const testArr = [20, 30, 40, 50, 80, 90];